Moved media rendering to presenter
